fix(sound-make): resolve float type mismatch in sample conversion
Casts the sample value to f64 explicitly during conversion to support f32 samples under the lite feature set.

+++ src/sound_make.rs
@@ -149,8 +149,9 @@ fn generate_wav(frequency: f32, duration: Duration, amplify: f32) -> Result<Vec<
     let sample_rate = source.sample_rate().get();
     let num_channels = source.channels().get();
 
+    #[allow(clippy::unnecessary_cast)]
     let samples: Vec<i16> = source
-        .map(|s| (s.clamp(-1.0, 1.0) * i16::MAX as f64).round() as i16)
+        .map(|s| ((s as f64).clamp(-1.0, 1.0) * i16::MAX as f64).round() as i16)
         .collect();
 
     let bits_per_sample = 16u16;
